not bad for a day
tweaked components, may need to remove rounding of the images being served as there is a silly white corner on the components.

Need to reanimate the gifs inside the cards

need to swap out skybox with relevant 360 image on hover of card.

Need to link from card to experience. (this is all in trello as well, obviously :+1: )